Updates to Compass Calibration and Compass Heading Calculation
- Add mutator to allow accelerometer to be added to a compass after construction
- Use floating point values for Sample3D:dSquared() operations to resmove possible overflow
- Change compass calibration hill climb algorithm to use 100 unit steps to converge faster.
- Update autodetect functions to register accelerometer instances with compass instances upon detection
- Clean and fix axis alignments for tilt compensation algorithm

+++ inc/drivers/MicroBitCompassCalibrator.h
@@ -106,7 +106,7 @@ class MicroBitCompassCalibrator
      *
      * @return The deviation between the closest and further point in the data array from the point given. 
      */
-    static int measureScore(Sample3D &c, Sample3D *data, int samples);
+    static float measureScore(Sample3D &c, Sample3D *data, int samples);
 
     /*
      * Performs an interative approximation (hill descent) algorithm to determine an
